move components into `/pages`
